Refactored code to align with feedback
